finish first pass on docs
