+++ spiel/__init__.py
@@ -2,7 +2,8 @@ from spiel.app import SuspendType, present
 from spiel.constants import __version__
 from spiel.deck import Deck
 from spiel.slide import Slide
-from spiel.transition import Direction, Swipe, Transition
+from spiel.transitions.protocol import Direction, Transition
+from spiel.transitions.swipe import Swipe
 from spiel.triggers import Triggers
 
 __all__ = [
